+++ src/api/tasks/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-task.dto';
+export * from './update-task.dto';
