+++ src/api/tasks/tasks.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { Task } from './entities';
 import { TasksService } from './tasks.service';
 import { TasksController } from './tasks.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Task } from './entities/task.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Task])],
